create custom hook for round list
